+++ colordialog.h
@@ -11,9 +11,9 @@ public:
 
     virtual Bool CreateLayout(void) = 0;
     virtual Bool InitValues(void) = 0;
-    virtual Bool Command(LONG id,const BaseContainer &msg) = 0;
-    virtual Bool CoreMessage(LONG id,const BaseContainer &msg) = 0;
-    virtual LONG Message(const BaseContainer& msg, BaseContainer& result) = 0;
+    virtual Bool Command(Int32 id,const BaseContainer &msg) = 0;
+    virtual Bool CoreMessage(Int32 id,const BaseContainer &msg) = 0;
+    virtual Int32 Message(const BaseContainer& msg, BaseContainer& result) = 0;
 
 	virtual void UpdateColor(Color color) = 0;
 };
