+++ colorpickerdialog.h
@@ -20,15 +20,15 @@ public:
 
     virtual Bool CreateLayout(void);
     virtual Bool InitValues(void);
-    virtual Bool Command(LONG id,const BaseContainer &msg);
-    virtual Bool CoreMessage(LONG id,const BaseContainer &msg);
-    virtual LONG Message(const BaseContainer& msg, BaseContainer& result);
+    virtual Bool Command(Int32 id,const BaseContainer &msg);
+    virtual Bool CoreMessage(Int32 id,const BaseContainer &msg);
+    virtual Int32 Message(const BaseContainer& msg, BaseContainer& result);
 
 	virtual void UpdateColor(Color color);
 	void FindICCProfiles();
-	void ChangeRGBSliderProfile(LONG index);
-	void ChangeCMYKSliderProfile(LONG index);
-	void LoadSpotColors(LONG index);
+	void ChangeRGBSliderProfile(Int32 index);
+	void ChangeCMYKSliderProfile(Int32 index);
+	void LoadSpotColors(Int32 index);
 
 	cmsHPROFILE m_displayProfile;
