+++ palettecolor.h
@@ -18,18 +18,18 @@ class PaletteColor :public SpotColor
 		~PaletteColor();
 
 		virtual Bool Init(void);
-		virtual void Sized(LONG w,LONG h);
-		virtual void DrawMsg(LONG x1,LONG y1,LONG x2,LONG y2, const BaseContainer &msg);
-		virtual LONG Message(const BaseContainer& msg, BaseContainer& result);
-		virtual Bool CoreMessage(LONG id, const BaseContainer& msg);
+		virtual void Sized(Int32 w,Int32 h);
+		virtual void DrawMsg(Int32 x1,Int32 y1,Int32 x2,Int32 y2, const BaseContainer &msg);
+		virtual Int32 Message(const BaseContainer& msg, BaseContainer& result);
+		virtual Bool CoreMessage(Int32 id, const BaseContainer& msg);
 		virtual void UpdateColor(Color color);
 		void UpdateBitmaps();
-		void SetColorID(LONG id){m_colorID = id;}
-		void SetPaletteID(LONG id){m_palette = id;}
+		void SetColorID(Int32 id){m_colorID = id;}
+		void SetPaletteID(Int32 id){m_palette = id;}
 		static void LoadIcons();
 		static void UnloadIcons();
 	private:
-		LONG m_palette, m_colorID;
+		Int32 m_palette, m_colorID;
 		HOVER_STATE m_hoverState;
 		static BaseBitmap *m_refreshIcon;
 		static BaseBitmap *m_leftArrowIcon;
