+++ spotcolor.h
@@ -14,9 +14,9 @@ class SpotColor : public GeUserArea
 		~SpotColor(void);
 
 		virtual Bool Init(void);
-		virtual Bool GetMinSize(LONG &w,LONG &h);
-		virtual void Sized(LONG w,LONG h);
-		virtual void DrawMsg(LONG x1,LONG y1,LONG x2,LONG y2, const BaseContainer &msg);
+		virtual Bool GetMinSize(Int32 &w,Int32 &h);
+		virtual void Sized(Int32 w,Int32 h);
+		virtual void DrawMsg(Int32 x1,Int32 y1,Int32 x2,Int32 y2, const BaseContainer &msg);
 		virtual Bool InputEvent(const BaseContainer &msg);
 
 		void SetParent(ColorDialog *parent);
@@ -28,6 +28,6 @@ class SpotColor : public GeUserArea
 		ColorDialog *m_parent;
 		Color m_color;
 		
-		LONG m_w;
-		LONG m_h;
+		Int32 m_w;
+		Int32 m_h;
 };
